+++ changelog.md
@@ -5,6 +5,7 @@
 - Add implicit reconcile to remove unknown Mesos Tasks
 - Add limitation of memory swap (DOCKER_MEMORY_SWAP, default value "1000" MB) the container can use
 - Add API and cli to shutdown the K8 cluster (does not delete persistence volumes)
+- Change kubeapi to a fix hostport. It would ever be the first port of the framework portrange
 
 ## v0.4.0